parseInt changed to parseFloat

+++ src/js/app.js
@@ -6,8 +6,8 @@ const resultEl = document.getElementById('result');
 const calculateEl = document.getElementById('calculate');
 
 function reactToClick () {
-    const fuelConsumption = parseInt(fuelConsumptionEl.value);
-    const amountOfFuel = parseInt(amountOfFuelEl.value);
+    const fuelConsumption = parseFloat(fuelConsumptionEl.value);
+    const amountOfFuel = parseFloat(amountOfFuelEl.value);
     resultEl.textContent = calculateDistance(fuelConsumption, amountOfFuel);
 }
